Set up for Heroku
Notably, `psycopg2` and/or postgres are picky about types when filtering in queries.
In particular, `Blogger.gh_id` is stored as a string in the database, while GitHub's
responses provide it as in int, so it has to be casted to a string.

We could take GitHub's word for it and put `int`s in the database, but depending on
data types from external sources feels sketchy