Forget trying to make subfolders work in place of subdomains
I can't figure out how to control which `URL` will be chosen by Flask/Werkzeug in case
more than one match. Flask tries to pick the match with the fewest variables, but here
I'd like to use different URLs with the same number of variables (`<subdomain>` becomes
part of the `rule`, like `/subdomains/<variable>/ + rest_of_rule` or something.

This would be great because while subdomains in production are awesome, they are a pain
to deal with locally. My current solution is to put lots of garbage in my `/etc/hosts`
file to send all the subdomains I might want to test with to `127.0.0.1`, and then
setting `export SERVER_NAME=commit--blog.local:5000` in my environment. This kind of
sucks. Lots to set up just to test the site.

Later I might try making a special `@route` decorator that will determine at run-time
whether to build a subdomain rule or a development subfolder rule.