Use git directly to fetch commit message
At long last, a first step toward supporting git in general and not just github.

When fetching the commit message and commit time, dulwich is now used to clone
the repo to memory and then grab the metadata. It's not efficient, but it works.

In the past I've used [libgit2](https://libgit2.org/) for programatic git
access, but this time I went for [dulwich](https://www.dulwich.io/). Dulwich is
a pure-python implementation, which makes both hacking on it and deployment
easier. Having the whole thing in a memory-safe language is also nice, although
with the heavy-hitters in git hosting relying on it, libgit2 is probably better-
reviewed.

Since cloning arbitrary repositories can consume arbitrary resources, the actual
clone runs in a `multiprocessing.Pool` with a 30-second timeout. If the time
limit is reached, it is retried with a shallow clone (depth 32) and a 20 second
timeout. I'm hoping that for now, this strikes an acceptable balance between
commit reachability and self-preservation.

Git hosts running recent versions may enable `allowReachableSHA1InWant`, or even
`allowUnreachable`, for fetching metadata by sha directly, like the github api.
This would save having to clone the whole repo, so this would be a good thing to
try first in the future before falling back on cloning.

Metadata-only MemoryRepo
------------------------

One simple optimization I included is to only keep `Commit` objects in the MemoryRepo when cloning. There is still a lot of processing that happens on the objects coming through, so this did not result in a speedup for large repositories, but it did cut the maximum resident set size significantly. A few quick measurements:

*Resident set size increases during fetch*

git@github.com:uniphil/feedwerk.git
 - vanilla:  2.19MB
 - treeless: 1.22MB

git@github.com:uniphil/rust-clippy.git
 - vanilla: 557.27MB
 - treeless: 75.09MB

Fancier timeout wrapping
------------------------

The code I went with is less complicated, but if I had to do more timelimiting,
I think this sketch was kind of nice

```py

def timelimit(t, fn, *args, retry=None, **kwargs):
    try:
        with Pool(processes=1) as pool:
            return pool.apply_async(fn, *args, **kwargs).get(timeout=t)
    except TimeoutError:
        if retry is None:
            raise
        t2, *args2 = retry
        with Pool(processes=1) as pool:
            return pool.apply_async(fn, *args2).get(timeout=t2)

if __name__ == '__main__':
    message = timelimit(30, get_message, (url, ref),
                 retry=(20, (url, ref), {'depth': RETRY_DEPTH}))
```