Use dulwich without subprocessing
Still unsure of the cuase of the hang with the previous configuration, but this
seems to work ok.

Unfortunately it drops the timeout, but I think gunicorn will still try to kill
it if a repo takes up more than the 30 seconds a worker has to respond.
Hopefully it doesn't get stuck after that point.

Tried adding `gevent` since these workers def should be async, but it takes a
while to compile from source, since this build is weird thanks to dulwich.

Thinking of using the Futures / coro threaded worker type instead. Need to read
more about it first though.