Fix #19: update readme.md
This is a first pass at an updated README, re: issue #19.

I didn’t go into depth for the **Contributing** section, since I’m still trying to figure that out. But once you update that, I will read it, and then start contributing in that way.

I usually like to call out the license from the README, but we can delete that part if it’s not relevant.

And speaking of deletion, please feel free to cut this up and rewrite it however you see fit! I hope it helps to have an initial draft in place, but that’s all this needs to be.

…can’t wait to commit--blog this 😂