Try out python-markdown for rendering
Opt in by adding `?rr=1` to the blog list, blog post, or rss feed endpoints.

If this seems to work, the next step will involve a database migration 😱. I dunno if i've ever done one with this project? 😅

My current thought is to add a `markdown_version` column to the posts table, so that posts rendered with old configurations are detectable. I _don't_ want to keep multiple copies / versions around, but I think it's useful not to re-render and potential change posts that were already blogged in the past.

When an author views an old post, we can offer to preview it with the current render config -- if they accept, it updates with no take-backs. But otherwise it just stays the same.

For a while I was thinking that maybe the markdown rendering could be pushed entirely out of the application entirely and into a front-end/middleware cache layer, but I... don't any more. Some nice things can happen if the as-posted render is treated as first-class data.