Preview and save markdown rerenders
Since the config for markdown rendering is likely to change in the future, this
adds detection of posts that have been rendered with previous configurations,
with the option to preview a rerender under the new config.

If the new render result is identical, the rerender and version bump is auto-
committed. If it differs, a preview is shown where it can either be committed
or left as-is.

Contains hacks; :shrug: