don't linkify in `code` bits
a tiny linkifier regression

`pre` is switched for `code` for places not to linkify, which should be ok
since the `codehilite` plugin always wraps code in `<code>` even inside the
`<pre>`. The only place I think this would be a problem is if someone used a
`<pre>` tag literally in their message and expected not to have it linkified.

🤷🏻 getting linkified isn't a huge deal hopefully, and it's easy to add `pre`
back if someone needs it.