Refreeze, add gevent, bye dulwich
The whole idea of using dulwich over libgit2 was to pick the pure-python (and
thus memory-safe) option for such a large, user-exposed feature. However,
getting dulwich to build without c extensions (--global-option="--pure") has
forces _all_ packages to build from source. With gevent now in the mix this has
just grown to too big a cost.

Additionally, it's looking more and more like this will need to grow into an
accidentally full-on git hosting service. Git is just not made for lightweight
crawling of metadata on remotes. Actually cloning whole repos and querying them
from storage is going to involve lots of reading and writing of files, which
feels like a much bigger security surface area to worry about, with memory
safety looking like a relatively smaller and smaller aspect.

All of this is basically pushing me back toward libgit2/pygit2 :(

It was fun for a bit