exclude prod-system packages from freezing
for packages that want to build source instead of using wheels (because openbsd), they tend to fail building (because openbsd). fortunately they are often available to `pkg_add` in openbsd, but pip will still try to install them to the venv. this change allows excluding files from `requirements.txt` that will be available as system-installed packages on the host instead.