Stop myself from reintroducing this bug again :'D
Diff exits non-zero when we just want to print the diff, so pipefail is going to give us bad times. I def looked at this and wondered "why didn't i add pipefail" as i put it in on the last change...

It could be disabled just around the pipe where it makes things break, but the extra noise is distracting and there aren't enough pipes in the whole script for it to be much benefit.